+++ CHANGELOG.md
@@ -5,6 +5,7 @@ See also [TODO](TODO.md)
 - [x] order save redirects to order details instead of order list
 - [x] order upload attachment available at any time, even if placed or filled
 - [x] order admin search box includes order ID
+- [x] fill admin search box searches by fill ID
 
 - [ ] add button "confirm" to order details which sends the order email to backoffice
   - no email should be sent at "save"